Fixed intake delay, changed drivetrain motor group
In order to isolate the drivetrain motors, I created a drivetrain function that actuates the PTO motors with the drivetrain only when the PTO is engaged. This removed all lag from actuating the intake or windback. The next thing that needs to be done is make the intake toggleable and the windback function (pullback until limit switch pressed/pullback until limit switch not pressed [firing catapult])